feat:add role_department logic

+++ spring2go-modules/spring2go-system-biz/src/main/java/com/spring2go/system/mapper/SysUserMapper.java
@@ -23,7 +23,7 @@ public interface SysUserMapper extends BaseMapper<SysUser> {
      * @param id 用户ID
      * @return userVo
      */
-    UserInfo getUserInfoById(Integer id);
+    UserInfo getUserInfoById(Long id);
 
     /**
      * 分页查询用户信息（含角色）
